Hack to fix the issue in multi node cluster, of not getting all mutations.
Added a new method to find all the nodes in a cluster.
Even if user doesn't provide all the nodes in the cluster, using
this method, we can find all the nodes in the cluster.

Updated the benchmark backup code to use this new method to find
all the nodes in cluster.
Connect to each node, one by one, and fetch the partition sequence
numbers on that node.
Merge all the sequence numbers fetched to get a full list.